+++ lib/vagrant-lxc/action/forward_ports.rb
@@ -72,8 +72,10 @@ module Vagrant
         end
 
         def redirect_port(host_ip, host_port, guest_ip, guest_port)
-          host_ip = "--laddr=#{host_ip} " if host_ip
-          redir_cmd = "redir #{host_ip}--lport=#{host_port} --caddr=#{guest_ip} --cport=#{guest_port} 2>/dev/null"
+          params = %W( --lport=#{host_port} --caddr=#{guest_ip} --cport=#{guest_port} )
+          params.unshift "--laddr=#{host_ip}" if host_ip
+          params << '--syslog' if ENV['REDIR_LOG']
+          redir_cmd = "redir #{params.join(' ')} 2>/dev/null"
 
           @logger.debug "Forwarding port with `#{redir_cmd}`"
           spawn redir_cmd